requirements/base.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/base.in -o requirements/base.txt
annotated-types==0.7.0
    # via pydantic
asgiref==3.8.1
    # via django
asttokens==2.4.1
    # via icecream
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
cheshire-cat-api==1.4.8
    # via -r requirements/base.in
colorama==0.4.6
    # via icecream
django==5.1.3
    # via -r requirements/base.in
django-widget-tweaks==1.5.0
    # via -r requirements/base.in
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via -r requirements/base.in
executing==2.1.0
    # via icecream
icecream==2.1.3
    # via -r requirements/base.in
idna==3.10
    # via
    #   email-validator
    #   requests
psycopg2-binary==2.9.10
    # via -r requirements/base.in
pydantic==2.10.1
    # via cheshire-cat-api
pydantic-core==2.27.1
    # via pydantic
pygments==2.18.0
    # via icecream
python-dateutil==2.9.0.post0
    # via cheshire-cat-api
python-decouple==3.8
    # via -r requirements/base.in
requests==2.32.3
    # via -r requirements/base.in
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sqlparse==0.5.2
    # via django
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.0.7
    # via
    #   cheshire-cat-api
    #   requests
websocket-client==1.8.0
    # via cheshire-cat-api
